fix(resource-download): 修复导航不生效问题 & 修复页面内容区外边距问题
